Tablas, posts, vistas del a plantilla, sql, Router
- Se renombraron las columnas:
> category_count -> category_post_count
> term_count -> term_post_count
> comment_status -> post_comment_status
> comment_count -> post_comment_count
- Nueva columna en la tabla users.
- Redireccion, al crear un post, a la pagina index.
- Corrección al establecer el usuario, creador del post.
- Se reenombraron variables y métodos.
- Ahora entes de borrar un usuario se comprueba si no es el mismo que el
de la sesión actual.
- Nuevas implementaciones en las clase Template.
- Nueva clase ThemeControllerAbstract.
- Correcciones al establecer los datos de la Ruta.
- Vista de la plantilla para los post, lista de post por usuario, lista
de post por categoría y lista de post por etiqueta.
- Actualizado Script SQL.

+++ app/controllers/admin/UserController.php
@@ -8,6 +8,7 @@ namespace SoftnCMS\controllers\admin;
 use SoftnCMS\controllers\CUDControllerAbstract;
 use SoftnCMS\controllers\ViewController;
 use SoftnCMS\models\CRUDManagerAbstract;
+use SoftnCMS\models\managers\LoginManager;
 use SoftnCMS\models\managers\UsersManager;
 use SoftnCMS\models\tables\User;
 use SoftnCMS\util\Arrays;
@@ -85,9 +86,11 @@ class UserController extends CUDControllerAbstract {
         $user->setUserRol(Arrays::get($inputs, UsersManager::USER_ROL));
         $user->setUserUrl(Arrays::get($inputs, UsersManager::USER_URL));
         $user->setUserPassword($pass);
+        $user->setUserPostCount(NULL);
         
         if (Form::submit(CRUDManagerAbstract::FORM_CREATE)) {
             $user->setUserRegistered(Util::dateNow());
+            $user->setUserPostCount(0);
         }
         
         return ['user' => $user];
@@ -175,17 +178,20 @@ class UserController extends CUDControllerAbstract {
     }
     
     public function delete($id) {
-        $messages     = 'Error al borrar el usuario.';
-        $typeMessage  = Messages::TYPE_DANGER;
-        $usersManager = new UsersManager();
+        $messages    = 'Error al borrar el usuario.';
+        $typeMessage = Messages::TYPE_DANGER;
         
-        $result = $usersManager->delete($id);
-        
-        if ($result === FALSE) {
-            $messages = 'No se puede borrar un usuario con entradas publicadas.';
-        } elseif ($result === 1) {
-            $typeMessage = Messages::TYPE_SUCCESS;
-            $messages    = 'Usuario borrado correctamente.';
+        if ($id != LoginManager::getSession()) {
+            $usersManager = new UsersManager();
+            
+            $result = $usersManager->delete($id);
+            
+            if ($result === FALSE) {
+                $messages = 'No se puede borrar un usuario con entradas publicadas.';
+            } elseif ($result === 1) {
+                $typeMessage = Messages::TYPE_SUCCESS;
+                $messages    = 'Usuario borrado correctamente.';
+            }
         }
         
         Messages::addMessage($messages, $typeMessage);
